chore: moved all elements from blocks to blocks/settingsBlocks
