+++ app/src/main/kotlin/view/appInterface/dialogElements/blocks/settingsBlocks/GeneralBlock.kt
@@ -1,4 +1,4 @@
-package view.appInterface.dialogElements.blocks
+package view.appInterface.dialogElements.blocks.settingsBlocks
 
 import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.foundation.layout.Column
@@ -24,9 +24,9 @@ import org.coremapx.app.userDirectory.config.ConfigKeys.IS_EXPANDED_SETTINGS
 import org.coremapx.app.userDirectory.config.ConfigKeys.LANGUAGE
 import org.coremapx.app.userDirectory.config.ConfigKeys.SYSTEM_DIALOG_THEME
 import org.coremapx.app.userDirectory.config.ConfigKeys.THEME
-import view.appInterface.preview.PreviewSurface
 import view.appInterface.dialogElements.lines.DropdownSelectLine
 import view.appInterface.dialogElements.lines.SwitchLine
+import view.appInterface.preview.PreviewSurface
 import java.util.Locale.getDefault
 
 @Suppress("ktlint:standard:function-naming")
