+++ app/src/main/kotlin/view/appInterface/dialogElements/blocks/settingsBlocks/WorkAreaBlock.kt
@@ -1,4 +1,4 @@
-package view.appInterface.dialogElements.blocks
+package view.appInterface.dialogElements.blocks.settingsBlocks
 
 import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.foundation.layout.Column
@@ -23,8 +23,8 @@ import org.coremapx.app.userDirectory.config.ConfigKeys.GRAPH_LAYOUT_HEIGHT
 import org.coremapx.app.userDirectory.config.ConfigKeys.GRAPH_LAYOUT_WIDTH
 import org.coremapx.app.userDirectory.config.ConfigKeys.VERTEX_LABEL_SIZE
 import org.coremapx.app.userDirectory.config.ConfigKeys.VERTEX_RADIUS
-import view.appInterface.preview.PreviewSurface
 import view.appInterface.dialogElements.lines.NumberTextFieldLine
+import view.appInterface.preview.PreviewSurface
 
 @Suppress("ktlint:standard:function-naming")
 @Composable
